requirements.txt
streamlit>=1.31.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
scipy>=1.10.0
plotly>=5.15.0
pyyaml>=5.4.1
openpyxl>=3.0.0
